+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - 2019-11-21 : (PLB2B-943)  : Build with Xcode 11.2
+- 2019-11-21 : (PLB2B-942)  : Provide extended BluetoothState including unauthorized state 
 
 ### Changed
 ### Fixed